Scan network from menu, add founded device on plus button

+++ app/src/main/java/pl/sviete/dom/devices/db/AisDeviceEntity.kt
@@ -4,7 +4,7 @@ import android.arch.persistence.room.*
 
 @Entity(tableName = "AisDevice")
 class AisDeviceEntity(
-    @PrimaryKey(autoGenerate = true) var uid: Int? = null,
+    @PrimaryKey(autoGenerate = true) var uid: Long? = null,
     @ColumnInfo(name = "name") var name: String,
     @ColumnInfo(name = "mac") var mac: String,
     @ColumnInfo(name = "ip") var ip: String?,
